backend/requirements.txt
fastapi[all]
sqlalchemy
pwdlib[argon2]
pyjwt
jwt
pydantic-settings
datetime
fastmcp
bcrypt
langchain
langchain_mcp_adapters
langgraph
langchain-community
langchain-core
ddgs
passlib
python-jose
trafilatura
tavily-python